read /proc/mtrr rather than use ioctl() interface
The /proc/mtrr ioctl() MTRRIOC_GET_ENTRY returns zero'd entries for
memory bases > 4GB which is wrong.  However reading /proc/mtrr
provides the correct entries for >4GB but we now have to manually
parse the output, which is a pain.

This patch removes the use of the ioctl() interface and parses the
/proc/mtrr output to get the all the MTRR entries.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by Alex Hung <alex.hung@canonical.com>