acpica: add -fno-strict-aliasing to ignore type punning warnings (LP: #971886)
We don't have much control over the quality of code in the ACPICA core that
is incorporated into fwts.  When compiled on Gentoo there are a bunch of
"dereferencing type-punned pointer will break strict-aliasing rules" warnings
which we can't easily fix.  For now, just use -fno-strict-aliasing on the
ACPICA core to workaround this noise.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>