lib: rework logging and framework to allow for multiple output logs
Since we now can output many different log types we should also
allow for multiple log types to be written during a run.  This involves
some considerable re-working of the logging engine.

1. The fw->log_type is now a  bit map of log_types
2. We add a list of log types to be written to fwts_log, this is a list
   of fwts_log_file types.
3. We need to re-work the log name handling so that we can open multiple
   log files with different suffixes depending on the log type.
4. To reduce the amount of vsnprintf() of the formatted log output we
   now handle this in the log fwts_log_printf() and pass down the formatted
   output to the different logging handlers rather than keep on re-formatting
   at the lowest logging handler layer.

There are a lot of changing is this patch. I tried to break it down, but
since there are so many interdependant changes I had to resort to one big
patch

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Tested-by: Chris Van Hoof <vanhoof@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>