lib: fwts_summary: remove references to log line number
With the changes to the logging back end, line numbers vary depending
on the output back-end being used.  So, referencing the log line number
is no longer trivial.  Removing the log line references in the summary
is probably the best way forward for the moment, otherwise we end up
writing a really overly complex log line tracking mechanism for something
that isn't too useful anyhow.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>