lib: fwts_acpi_tables: load tables in deterministically
When loading a bunch of ACPI tables we should load them in
some form of deterministic order rather that the random order
that they appear in the directory.  So, load them in on
filename alphasorted order.  This allows us to sanely check the
tables using fwts-test with some sort of determinism.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>