lib: fwts_acpi_tables: add ACPI table provenance
It is useful (for example in the acpidump test) to be able
to determine where an ACPI table came from.  In normal mode, fwts
loads ACPI tables directly from firmware in memory, but it can also
load them from file (e.g. from acpidump) or even generate them if
they are missing in a fix-up processing stage.

So, lets add an extra field in the acpi table structure that keeps
track of each tables provenance.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>