apci: acpitables: only check CMOS in BOOT table from firmware tables (LP: #1016469)
The fwts acpidump test currently always checks the CMOS settings on BOOT
tables no matter where the tables were loaded from. The CMOS settings
validation is only valid when loading the tables directly from firmware.

To be able to check the table provenance we need to pass over the acpi table
struct to the BOOT test to check this flag.  Unfortunately this means we need
to alter all the callback functions so we can pass the table struct rather than
the table data and the table size.  This means we also should fix the table sizes
in the code from int to size_t to ensure consistency.

This patch therefore changes:

1.  The prototype of all callback table checking functions to just table
    acpi tables rather than table data and length.
2.  Fix the table lengths in called code from int to size_t.
3.  Check the CMOS in the BOOT table ONLY when the table is loaded
    directly from firmware.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>