lib: fwts_hwinfo: sort device info before comparing (LP: #1018288)
Output from xinput list can sometimes be re-ordered beween and
suspend and a resume, leading to fwts complaining that there are
differences between input devices even though it is all OK. The
simple workaround is to sort the output alphanumerically first
and then compare.

We also filter out the pactl Latency information as this is
a new field that we don't want to compare because it is always
changing.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Tested-by: Chris Van Hoof <vanhoof@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>