acpi: acpidump: dump out ACPI UEFI tables
Dump out the ACPI UEFI tables, mainly just the GUID and the raw
hex dump of the raw data.  We also add acpi_dump_guid() to allow
us to format up GUIDs and also add an offset argument to the raw
data hex dump to allow us to dump out sections of a table in hex
while printing the table offset address.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>