lib: fwts_acpi_tables: Force fixup on XSDT and RSDT when loading from .dat files
When loading raw ACPI tables from .dat files we have no idea where the original
.dat tables were in memory, so this means that loading in the XSDT and RSDT is
problematic because these contain references to the original tables in memory
but we have no clue how these match up with the raw data.  Since fwts has to
fake the addresses of the loaded ACPI tables these don't match those in the
loaded XSDT and RSDT which leads to the ACPICA core getting confused.  The
best strategy is to therefore not load the XSDT and RSDT and rely on the fwts
table fix up stage to notice we haven't loaded the XSDT and RSDT and let this
create these tables based on the tables we have loaded from the raw .dat data,
leading us to having sane XSDT and RSDT that correctly reference the
loaded ACPI tables even if they are at faked addresses.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>