lib: acpica: fix semaphore counting by waiting for threads to complete
When executing Notify() handlers ACPICA creates separate threads which
we need to wait to complete before we do any semaphore usage accounting.
This is because these threads can themselves use internal or AML semaphores
which will be in an unknown state if we do the start doing semaphore
usage accounted before we wait for them to terminate.  This patch adds
in an array to keep track of threads created by AcpiOsExecute and also
a pthread join on these running threads before we start counting any
used semaphores.

Note that we have to deprecate the ACPICA verison of AcpiOsExecute and
re-implement this in fwts to allow us to do the thread creation tracking.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu<ivan.hu@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>