lib: fwts_klog: sanity check json klog data is readable
Unfortunately the json library is quite flakey and sometimes
loading the klog json data from a non-existant file causes the
json library to segfault.  This patch is a workaround that
sanity checks to see if the data file can be opened for reads
before attempting to call json_object_from_file().

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>