uefi: add in support for new uefivar file system interface
Currently we read uefi variables via a /sys interface, however a
new kernel feature will export these variables via a psuedo file
system, so add support for this.

This is rather a large change.  We abstract out the reading of
uefi variables using two new helper functions:

  fwts_uefi_get_variable_names() and
  fwts_uefi_free_variable_names().

We modify the abstraction of the UEFI variables, before we used
to have a 1-to-1 mapping onto sysfs, now we have to cater for large
sized variable contents, so we now make the data size and variable
name size allocated on the heap.  We also add in support for the
new efifs and abstract this and the /sys variable reading into
two helper functions.

Finally, this patch modifies the uefidump test to use these
changes in the uefi API.
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>