lib: framework: Add "unsafe" test category
Some tests we may add in the future (such as UEFI run time services)
can cause the kernel to oops if the firmware is buggy, so we should
add a new "unsafe" test category for these kinds of tests. Currently
there are no "unsafe" tests in fwts, so this does not make any difference
to output to fwts --show-tests until we start adding the actual tests,
so this is a fairly benign addition.

Also add the --unsafe and -U option to run unsafe tests.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>