lib + tests: bulk rename of fwts flag macros
The FWTS_FRAMEWORK_FLAGS_ macros were originally really rather too long
to type, so there were a bunch of #defines to shorten these for common
use in just the fwts tests.  However, the aliasing of these is really not
a good idea after all as it leads to confusing mixes of the same value
with different names, especially in the fwts framework, for example

FWTS_BATCH is really an alias of FWTS_FRAMEWORK_FLAGS_BATCH (urgh!)

This patch simplifies the matter by renaming all these macros from
FWTS_FRAME_WORK_FLAGS* to FWTS_FLAG_* and we replace the bad short cut
versions to the FWTS_FLAG_* equivalent.

This changes no functionality in the code whatsoever, it just tidies
it up and makes it more understandable.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>