lib: fwts_json: wrapper for json error returns
Older versions of json-c may return an error in an
object as a ((json_object*)-1), where as newer
versions return NULL.  To cope with this API change
we create a macro FWTS_JSON_ERROR that will return
true if an object pointer is flagging an error
condition as a NULL or a ((json_object*)-1) pointer.
This way we can cope with errors no matter which
version of json-c we use.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>