klog: change the MSR_IA32_ENERGY_PERF_BIAS test from fauilure to info
Intel IA32 program guild doesn't mention that this MSR should be set by
firmare, instead, it mentions "Software can use whatever criteria it sees
fit to program the MSR with the appropriate value.". Discuss with IBV (AMI),
they'd like to keep this MSR as default, and let OS to change value if desired.

Signed-off-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Colin Ian King <colin.king@canonical.com>