lib: fwts: add FWTS_UNUSED() macro for unused func arguments
I'm hoping to use -Wextra in the near future to ensure code quality
is high.  Some of the fwts functions such as callbacks and fwts library
helper functions take arguments that we ignore.  To hint to gcc that
we're really ignoring these and to stop -Wextra breaking I'm adding
FWTS_UNUSED() to explicitly tell us that the argument is meant to be
ignored.

Example:

void fwts_somefunc(fwts_framework *fw, int dummy) {
	/* variable declarations here */

	FWTS_IGNORED(dummy);

	/* start of executable statements here */
	..
}

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>