acpi: pcc: New test to sanity check PCC, (LP: #863175)
This adds a test to sanity check the Processor Clocking Control (PCC).
PCC seems to be found on just a few HP Proliant machines at the moment,
for example:

  HP ProLiant DL380 G6
  HP ProLiant DL380 G7
  HP ProLiant BL460c G7
  HP ProLiant BL460c G7

and probably a bunch more HP kit besides.  I've only found 4
different machines that have the PCCH out of ~6000 kernel related
bugs filed in LaunchPad, so this really is for a small handful
of machines.

This test evaluates PCCH and sanity checks the data structures.
We can go a bit further and determine the PCC Header and memory
map this and sanity check this too, however, for this initial
version I have disabled this functionality because I've not got
any hardware to test this on and also I am concerned that reading
an incorrectly specified header region may create some issues.
I will enable this once I can test this on real hardware.

For more information on PCC, see:
http://acpica.org/download/Processor-Clocking-Control-v1p0.pdf

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>