acpi: acpiinfo: Complete reworking to bring it up to date.
The apciinfo test is a legacy test originally derived from the
Linux firmware developer kit and had a lot of hard-coded kernel
log checks that are now either out of data or nor used.  I've now
moved these kernel log checks into the fwts klog table because it
makes more sense for them to be scanned by the klog tests.  I have
also added the recent kernel versions of these kernel log messages
to keep it in sync with the upstream code.

I was thinking of removing this test, but the acpiinfo name lends
itself to dumping out some ACPI specific information that is useful
to know.  So I've re-written the test to determine the ACPI version
supported by the kernel ACPI driver, the ACPI version in the
machine's firmware and I've completely re-worked the original code
to also determine the version of the ACPI compiler used to generate
the DSDT and SSDT AML.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>