lib, tests: remove fwts common test, merge into fwts klog
The fwts common test is really a tiny subset of the klog test. It
was included from the historical Linux firmware developers kit but
now makes no sense to keep on its own.  This patch removes the common
helper code from fwts_klog.c and from the s3 and s4 tests.  The
common test data in klog.json is now merged with the klog firmware
patterns and the fwts common tests is removed.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>