lib + tests: modify fwts_pipe_exec to return the child exit status
Fix bug that has been in the fwts s3, s4 and s3power tests. Originally the
semantics to fwts_pipe_exec() was that was meant to return the child exit status
but this got changed to return FWTS_OK if successful and FWTS_EXEC_FAILED if
the child exec failed.  Thus we no longer passed back the child exit status which
the s3, s4 and s4power tests relied on to check the suspend or hibernate pm script
exit status.

This patch adds an extra child exit status parameter to fwts_pipe_exec() which can
be checked if required.  However, for most use cases, one can ignore this and just
check for FWTS_OK or FWTS_EXEC_FAILED if you don't care about the child exit status.

The patch fixes up all the calls to fwts_pipe_exec and also changes the way errors
are detected assuming anything that is not a FWTS_OK return is a failure.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>