lib: cpu: add common CPU performance measuring code to fwts library
The C-states and CPU frequency tests use fairly common code to load
the CPUs.  Move this into fwts_cpu.c as fwts_cpu_performance() which
measures the performance of a specified CPU for just 1 second and
returns a bogo metric of CPU performance.  Also add in a CPU load
helper function.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>