hpet: hpetcheck: add ACPI HPET table check
The original test has some notion of checking the HPET table
but this was #ifdef'd out from the original Linux Firmware
Developer Kit code and never implemented in fwts.  Remove the
old legacy code and fully implement a HPET table validation.

Since we want to sanity check where the kernel's view of where
the HPET is located and what the HPET table states, I've re-ordered
the sub-tests to ensure the new test runs 2nd.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>