debian: change the mail address of the maintainer
Since the main mailing list is the most used one, no mail gets lost
or unattended if it goes there.

Further Ubuntu dpkg-dev checks this field if there is any distro fix.

If the package version has the suffix like -1ubuntu1, the mail needs
to be @ubuntu.com. If not, dpkg-source reports an error.

Signed-off-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>