lib + tests: don't stringify the test name in FWTS_REGISTER
FWTS_REGISTER currently takes the test name and use this as the
init function stub and also stringifies this to provide the
literal string name of the test.   This makes the macro a little
confusing since test name should be passes as a literal string
and we don't really need to add the name of test to the init
function stub since these are static declarations.

This patch removes the stringification and also modifies all
the tests to now pass in a literal string.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>