lib: fwts_framework: assert on FWTS_REGISTER names being too long
We don't want tests registered with stupidly long names so add in
the FWTS_ASSERT macro that can be used to throw a compile time
error if the name of a given test with FWTS_REGISTER is too long.

The FWTS_ASSERT macro is generic enough that we can use it for
other compile time sanity checks if required.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>