acpica: fwts_acpica: force indefinite AML Wait() to timeout
Quite a few _EJ0 implementations make the ACPI method test block
forever because they are using an AML Wait() that is waiting
indefinitely for an external event (which fwts cannot generate).
Instead of the test waiting forever for the event, make these
waits timeout after 20 seconds.  We force this by causing an alarm
signal which interrupts the blocked sem_wait and we can detect this
and report back to the user that we aborted the wait.  This is ugly
but effective.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>