acpi: method: tidy up passed/failure messages
Many of the ACPI method tests use the short name of the
control or object rather than the full path name.  We should
print out the full path name to clearly identify the specific
ACPI object path so we can see exactly which one is causing a
problem.

This patch also uses some more of the passed message helper
functions to as defined in an earlier patch.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>