acpi: method: Add helpers to check for package sizes
We do a lot of package size checking, so add two helper functions
to check for the minimum packake size required and also for the
exact package size required.

We can then use these and bail out of a test early of the package
size test fails.  This reduces the amount of code and also means
we can tidy up some of the complex nested if statements.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>