lib: framework: Increase precision of progress, add a spinner progress indicator
Some tests such as the cpufreq and cstates tests take rather a long time so it
makes sense to increase the precision to the % progress output to indicate that
something is actually happening.  Also, add a progress spinner indicator to also
provide some feedback that the test is actually active rather than hung.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>