acpi: method: add helper to check for package contents, simplify code.
Add method_package_elements_all_type() that checks to see if the
package contains elements that are all of a specified type and
method_package_elements_type() that checks the types against an
array of givem types.

These two helper functions allow us to remove a lot of duplicated code
and remove the need to do a lot of the error prone per object type
checking.

Also replace all "int failed" vars with booleans to simplify the code.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>