acpi: method: Add better _CRS resource checking
We're occasionally seeing some firmware with invalid or buggy
_CRS buffers being passed to the kernel and ACPI 5.0 has added
some new _CRS features so it seems prudent to sanity check the
buffers a little more.

This patch adds some _CRS sanity checking for small and large
resource types as defined by the ACPI 5.0 specification. I've run
this against my database of ACPI tables and it does seem to catch
a few minor and obscure _CRS issues.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>