script: smatchify.sh: build fwts and check with smatch
Add a bash script to automatically fetch and build the latest version
of smatch and then build fwts with this to find any coding errors.

This may find some false positives, but smatch is very useful at finding
some more subtle bugs so I think the false positives are worth the pain.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>