kernel: version: gather kernel boot command line
We're seeing some logs where fwts is being used with BIOS or ACPI
options being tweaked by users and it would be very helpful if fwts
can gather this information in the log.

Add this to the kernel version information gathering test, insert it
before the kernel ACPI version subtest.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>