acpi: wmi: re-write to eval _WDG rather than parse AML
This patch is a re-write of the acpi WMI test.  It no longer
parses the disassembled AML but instead evaluates all the _WDG
objects and parses and dumps the associated buffer.  This is a
little slower but should remove all the complexity of parsing
disassembled code which could break if the output changes in
the future.

Also, do some sanity checking to see if methods associated
with WMI events exist or not.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>