uefirtvariable: Correct DataSize for GetVariable()
When calling GetVariable(), we need to fill out DataSize with the size
of the Data buffer instead of using whatever garbage value is left on
the stack.

Furthermore, the Data buffers are not strings in any of the tests, so
there's no need to NUL-terminate them or make them 1-byte larger than
datasize, e.g. there's no real need to do this,

    uint8_t data[datasize+1;

since the last element of data[] is never read back anyway.

Cc: Ivan Hu <ivan.hu@canonical.com>
Cc: Colin Ian King <colin.king@canonical.com>
Cc: Alex Hung <alex.hung@canonical.com>
Signed-off-by: Matt Fleming <matt.fleming@intel.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>