acpi: syntaxcheck: re-work IASL error mappings
The IASL error codes may possibly change from release to release
of acpica so include the error message #defines from the compiler
and stringify these into a map.  We can then load the advice based
on the strifigied ids rather than the (changeable) error code. This
way the advice maps to the error message #defines rather than on a
changeable error code.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>