lib: framework: add in test name before test description
As suggested by John Brier, it makes sense to include the name
of the test, so lets prefix that before the test description.
This allows users to find the test results in the output log
by just looking for the test name.

An example of the output goes from:

Gather BIOS DMI information.

to:

bios_info: Gather BIOS DMI information.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>