uefirtvariable: Check new VariableNameSize from GetNextVariableName()
Some firmware implementations update VariableNameSize in
GetNextVariableName() with a value that is larger than the actual
buffer required to hold the VariableName string. This is not
technically a bug, but most implementations either update
VariableNameSize with the value of strlen(VariableName) + 1 or leave
it as the initial value passed to GetNextVariableName().

Print a warning if a different value is found.

Signed-off-by: Matt Fleming <matt.fleming@intel.com>
Acked-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>