cpu: cpufreq: make test more generic for non-x86 systems
Make cpufreq tests more generic and remove x86 only restriction
for non-x86 platforms.  We need to remove some of the x86 test
phases and re-phrase some of the output messages for the non-x86
case, but we end up with a test that is more generic.  Tested
on an ARM Nexus 4 phone.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>