cpu: microcode: rewrite so we no longer need to load microcode (LP: #1120240)
Rewrite the microcode test so that we no longer have to load the
microcode.  The microcode loader fails to work on recent releases
and re-writing/updating the microcode loader is required to make this
work again

Instead, just scan the kernel for microcode update messages and
sanity check the CPU microcode version.

Tested on Lucid through to Suacy releases.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>