lib: fix assembly/disassembly segfault when AcpiOsVprintf is called (LP: #1195209)
fwts --dis or fwts syntaxcheck can segfault if the ACPICA disassemble emits a
message via AcpiOsVprintf(). This calls the fwts_acpica helper function that
attempts to print to the log file which fails because the fw pointer is null.
This bug is a regression introduced with commit
54f1ef23a1c3bcbea023ddc799dc7d6becf60556.

To fix this, add a helper function to fwts_acpica that sets up the fwts
framework pointer so that fwts_acpica_vprintf won't segfault. Also sanity
check the framework pointer in fwts_acpica_vprintf in case it isn't set up
correctly.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>