lib: add acpica mode types
These need to be in their own header file because we can't put them
in fwts_acpica.h because this includes fwts_acpica.h and this requires
fwts_framework to be defined which needs acpica mode types defined
earlier. So, put them in their own header file to work around this
interdependency mess.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>>
Acked-by: Alex Hung <alex.hung@canonical.com>