fwts_iasl_interface.c: ensure we read in all of stdout (LP: #1200426)
Redirecting stdout to a file when running the syntaxcheck test results
in no failures.  We need to ensure we fflush on stdout to flush the
IASL output on stdout down the pipe to the reading parent process.

This patch also ensures we check for NULL on realloc.  Also fixes
an incorrect stdout fd check on pipefds[0] and removes an unnecessary
close on pipefds[1] on the child process.

Took the opportunity to do some white space cleanup and a general tidy
up of the code too.  Must of had a bad day when I originally wrote
this code.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>