acpi: syntaxcheck: use ASL_MESSAGE_TYPES instead of #defining them (LP: #1200568)
The IASL error codes are hard coded in fwts syntaxcheck and now differ
from the ones in ACPICA, so we are no longer picking up errors.
Instead we should use the ACPCIA error codes as defined in aslmessages.h
which we are already #including by are subsequently redefining. So just
remove the dodgy #defines and it all works correctly.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>