lib: fwts_log_json: check sting from json is not null (LP: #1206285)
fwts_log_close_json() assumes json_object_to_string() always succeeds.
It may not and this will lead to a segfault, so check for null and
don't bail out with an exit but continue on because user may be using
other logging back-ends and some output is better than aborted output.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>