acpi: fadt: sanity check reset register settings (LP: #1230394)
Probably a good idea to test for ACPI compliance on the reset register.
This extra test checks that the reset register width is 8 bits and the
offset is zero, as per the requirement of section 5.2.9 of the ACPI 5.0
specification.

The patch also shuffles up the fadt_size check to the initialization
section and fixes checks for FADT headers > ACPI version 1.0 in
test 1.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>