acpi: brightness: break up brightness test (LP: #1250429)
CERT were asking if the brightness test could be run automatically.
Tests 1 and 2 of the brightness test don't need user interaction
and hence should not be in an interactive test and these tests also
meet CERTs requirement.

This patch adds in brightness-helper for common brightness tests
helper functions.  Tests 1 and 2 of the brightness test now go into
a new autobrightness test and tests 3,4 and 5 are renamed and kept
in the original brightness test.

The patch set also tidies up the original scruffy code to remove
a stupid amount of nesting.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>