acpi: syntaxcheck: non-x86 may have ACPI support (LP: #1251293)
Working on the assumption that ARM based systems will be having ACPI
we may as well enable this for non-x86 architectures.  Anyhow, one
can also load ACPI tables into fwts and test them using this test on
non-ACPI systems so we may has well include this test just on that
reason too.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>