Allow RSDP to be provided manually or scan for it in klog (LP: #1260400)
During work on an early development machine we found a situation where
the RSDP could not be automatically probed using the current methods.
So, allow the RSDP to be specified either manually using the new -R or
--rsdp options or found by scanning the kernel log (ugly, but is our
final easy option).

We should also sanity check the RDSP by ensuring it lies on a 16 byte
boundary and contains a valid signature.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>