updated automake files for libjson-c (LP: #1268005)
This is a re-working of the earlier patch from Jeffrey Bastian. I've
had to modify it so that it does not break builds on earlier releases.

I've removed json lib references from the Makefiles and added an
configure.ac check to find the appropriate library and this adds it
to $LIB.  Tested the build in precise, saucy and trusty amd64 chroots.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>