lib: fwts_acpi_tables: fix parsing of empty data (LP: #1281197)
Acpi data as generated by the acpidump tool or fwts --dump with zero
data on a line can cause the parser to segfault if the data offset
is legitimate and the data isn't and the data is the first entry in the
table.  This occurred on some test tables that had got corrupted. The
fix is to break when we find no legitimate data.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>