acpi: fwts_acpi_tables: fix up missing FACS (LP: #1284097)
If fwts loads ACPI tables from file that don't have an FACS
we get all sorts of weird issues, for example, we are missing
the global lock. The best bet is to fix this table up by hand
and warn about this rather than cause all sorts of false
positives on the method tests.

This patch also re-works the table fixup to ensure we fixup
with the correct fixed-up table addresses for any tables that
we have to create because they are missing.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>