lib: fwts_acpi_tables: handle zero 64 bit addrs in FADT (LP: #1285167)
Some recent firmware has a FADT that is >= 140 bytes with zero 64 bit
X_FIRMWARE_CTRL and a non-zero 32 bit FIRMWARE_CTRL.  These kind of
errors need to be detected and fixed up rather than silently failing.

This fix reworks the FADT handling and adds a lot more error detection
and warning/error messages.  If an error occurs when loading the DSDT
or FACS we now clear up all the ACPI tables so that acpica initialisation
failes and we can bail out of tests in the init phase rather than causing
ACPICA to break in unpredictable ways.

The patch also modifies the acpi table load state so we can now detect
if the tables were previously loaded but failed, so we can avoid trying
to re-load the broken tables multiple times when we access tables in
various tests.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>