lib: fwts_acpi_tables: update RSDP checksum on fixed-up RSDP (LP: #1287852)
When loading the RSDP from a file sometimes the XSDT and/or RSDT
have been updated (fixed up) which requires clearing the old
checksums and re-calculating them for the updated RSDP. Currently
this is not occurring, causing fwts checksum tests to fail on the
RSDP on fixed up tables loaded from file. This patch fixes this.

Signed-off-by: Colin Ian King <colin.king@canonical.com>
Acked-by: Ivan Hu <ivan.hu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>